Streamlit Real-time API-based/requirements.txt
streamlit
pandas
python-dotenv
plotly